Initial Commit
Currently, the only sorting algorithm that has been implemented is the Bubble Sort. The current functionality only includes user input regarding the length of an array, which will then be used to generate an array and run it through the bubble sorting algorithm. The next major step will likely be the development of the GUI, and then more sorting algorithms.