GUI Created
Overall, this commit is heavily focused on the introduction of the GUI. Currently, it uses three separate JPanel objects within a JFrame object to represent the three different main parts of the GUI: The input, the main panel, and the output. The GUI is functional in that it can take in inputs, and use the user's given sorting algorithm to print the sorted version to the console. The next step is to finalize the input panel of the GUI, and then introduce the main and output panels. The following are specific changes made in this commit:

1. Added setter/getter for the elements field in SortingArray

2. Changed the built-in range for the array random number generation from [0,100) to [1,500).

3. The chooseSortingMethod method was no longer needed as a result of the GUI, so it was removed.

4. The setAttributes method was renamed to generateValues and now strictly generates the numbers for the input array.

5. When the program is first run, the main method in SortingArray creates a SortingArray object and a SortingMethod object called 'sorter'. When sorter is created, a constructor will create a hashmap with references to the various sorting methods that have been implemented thus far. These can be referenced by the GUI later on when the user presses the 'sort' button. After the hashmap is initialized, the input array used in the GUI is passed to the curArray field in SortingMethod for future use.

6.  Bubble and Reverse Bubble sort, the two sorting methods currently implemented, have been changed from subclasses to methods of SortingMethod.

7. New method for SortingMethod, called selectMethod, which will take the user's choice from the GUI, look in the previously mentioned hashmap for the correct sorting method, and then run it. This is done with lambda functions in the hashmap.