GUI Completion
The GUI has been completed. It is completely functional and capable of displaying sorting algorithms as they move through a generated array.

Currently, the speed at which this is done is hardcoded. This will be changed in a future commit. Of course, more sorting algorithms will also be added as well as further quality of life improvements to the GUI itself and backend optimizations.